Update README with Toy Robot brief
Update the Toy Robot README documentation with the application brief, in
order to describe the application functionality, purpose and criteria
that it needs to deliver. This brief describes the expected commands
that Robot is to received and response to, as well as provide examples
of expected output that can be later used for testing purposes.

== References:
- [Stack Exchange - Code Review - Robot toy simulator]
  (http://codereview.stackexchange.com/questions/96181/robot-toy-simulator)