Add Point class to represent cartesian points
Add a Point class to represent a cartesian point that the Toy Robot will
later used to navigate its simulated environment of Points within a grid
(such as the Table mentioned in the brief/README). This Point class is
based of the initial design draft with its primary functions to report
back is current coordinate as a simple string; e.g. "0,1"; and to
return a new Point with the sum of two Points added together, in order
to later update the Robot position, but without mutated the Grid
(effectively making it a Value object).

== Notes:
- Add appropriate associated tests, to ensure the Point class can be
  required via the Toy Robot library, can be constructed and works as
  expected

- Update `toy_robot` script to require the new Point class

- Remove unnecessary failing tests from Toy Robot library class, in
  order to green passing tests for all tests