Update Robot test to hide REPORT output
Update the Robot tests related to testing the `report` output method,
which outputs directly to STDOUT and appears within the unit test
output; e.g.

```
    $ rake

    Run options: --seed 25666

    # Running:

    ............................................"0,0,NORTH"
    ......

    Finished in 1.315907s, 37.9966 runs/s, 37.9966 assertions/s.

    50 runs, 50 assertions, 0 failures, 0 errors, 0 skips
```

This replaces the global `$stdout` variable temporarily to filter output
this console `puts` command and ensures that it is reset to the original
`STDOUT` once the test has finished (or in case it fails).

== Notes:
- Alternatively the Robot could be given an output (and possible Input)
  object to send its reports out to, rather than logging its report
  directly to standard out `$stdout`, via the `puts` method

== References:
- [Simple way to test IO in Ruby]
  (http://dalibornasevic.com/posts/39-simple-way-to-test-io-in-ruby)

- [Episode #455: Ensure - RubyTapas]
  (https://www.rubytapas.com/2016/11/16/episode-455-ensure/)