Add Toy Robot application binstub `bin/app`
Add an application bin stub for the Toy Robot library, which sets up the
Robot object and passes it input from either User input, after executing
the script, or the file loaded from the first argument passed into the
script (otherwise it will default to User Input if no arguments). It
will then loop over the input until either the `EXIT` command is given
or the end of file has been reached, depending on what input is given.

This successfully completes the Toy Robot specification, but can
definitively be improved to make testing easier and more isolate/robust,
and because this implementation places the processing of input outside
the application library (which is debatable if it should live as part
of the Toy Robot, since it implements its own commands independent of
outside input).

== Notes:
- Update READE document to reflect usage changes to the application,
  with some examples (which are also included within the test suite)

- Add some tests for evaluating the binstub as is, but also provide an
  possible future refactoring for the command line interface in order to
  better test and isolate the core loop of the application

- Add several file fixture examples, based directly off the Toy Robot
  specification

- One test in particular reflects the contents of the README, including
  the initial list of Commands that the Robot needs to impement, hence
  the more than expected output in its test; i.e.

  `def test_bin_app_returns_expected_output_for_readme_file`

== References:
- [Simple way to test IO in Ruby]
  (http://daliborneasevic.com/posts/39-simple-way-to-test-io-in-ruby)

- [For testing io stdin and stdout interaction. - GitHub]
  (https://gist.github.com/martinos/4660071#file-local_io-rb)