Refactor App binstub Input into classes
Refactor the possible File and _"Player"_ Input options into their
classes so that it can process the given input independent of its
source's `gets` method. This helps simpley the application binstub,
removing a lot of its responsiblity, which extracting and isolating the
majority of its logic into a more appropriate class.

These Input classes are loosely based off Tomas Varaneckas
implementation of its `PlayerInput` and `AIInput` components from their
"Developing Games With Ruby", which passes the game object's or entity's
control over to the Input class, but only after it has been initialized
in order to avoid the _Chicken and Egg_ problem with dependency
initialization order. It also means that potentially in the future the
Input for a given Entity can be switched out with another Input at
runtime (should that ever need to happen).

Also, as a part of this Input refactor this refactors the _Game_ or
_Simulation_ loop of the application to a single line, which only cares
about `update` from the Input, since it is the only component that
updates (the Robot itself updates its position or facing direction, but
it is directly dependent on the User / File Input, and does not need to
happen asynchronous to the processing of input or simulation, since the
application is "turn-based"; as in it will wait for input and only
update after recieving input).

== Notes:
- Add the appropriate tests, for both the abstract class `Input` and its
  subclasses, `PlayerInput` and `FileInput`

- Refactor existing instances of the process Input method, such as the
  one used in the test class `CLI` (which could be used to replace the
  application bin stub, fairly soon)

- Use inline test mock objects to fill in for different objects used
  within the Input class, so that they may tested in isolation of other
  library components

- The `source` method was not implmented within the Abstract/Base Input
  class, since it already raises the `NoMethodError` exception, which
  accomplishes its abstract nature, however it does implicitly avoid the
  `NotImplementedError` exception, which is more unimplement OS libraris
  rather than internal class methods

- The Input classes use a simple version of the Template pattern with
  its `source` method, alternative the source value could be passed in;
  e.g. `File.new(ARGV.first)` or `$stdin`; but these hold onto that
  responsiblity, rather than the application binstub instead

- Fix a slight spacing issue with one of the Robot tests `Point`
  initialize call

- A typical Game / Simulation loop looks like:

  ```
    while (true)
    {
      processInput();
      update();
      render();
    }
  ```

- The File and Player Input sub classes have been kept within the Input
  script for simplicity (for now)

== References:
- [Class: IO (Ruby 2.3.1)]
  (https://ruby-doc.org/core-2.3.1/IO.html#method-i-gets)

- [Read Developing Games With Ruby | Leanpub]
  (https://leanpub.com/developing-games-with-ruby/read#leanpub-auto-implementing-ai-input)

- [Please Stop Abusing NotImplementedError]
  (http://chrisstump.online/2016/03/23/stop-abusing-notimplementederror/)

- [Class: Module (Ruby 2.1.0)]
  (http://ruby-doc.org/core-2.1.0/Module.html#method-i-alias_method)