Add Grid class, to describe table world later on
Add Grid class to the Toy Robot library in order to describe the minimum
and maximum dimensions of the world that the Robot will inhabit in order
to separate that responsible from the current Robot, which can only
exist in a 5x5 grid.

This Grid class loosely describes the table mentioned in the Toy Robot
specification: i.e.

> - The application is a simulation of a toy robot moving on a square
>   tabletop, of dimensions 5 units x 5 units.
> - There are no other obstructions on the table surface.

This Grid describes the Table via its maximum and minimum units that
will be later used to restrict the movement of the Robot to only valid
moves within this context.