Update Robot to initialize itself within the Grid
Update the Robot to initialize itself with a reference to the world it
inhabits, and in the case of this application it is the `table`, which
is an instance of the Grid class. Since it now inhabits this new world
it can now evaluate its next move within the context said world, to
ensure that the next move it will take is indeed valid.

== Notes:
- Add a mock grid within Robot tests, to ensure it is not completely
  dependent on the Grid implementation as is

- This `valid_move?` expression can be moved out of the Robot entity and
  into the Grid world instead, so rather than the Robot knowing about
  the minimum and maximum dimensions of the world it inhabits, it can
  then instead ask the world if its next move is valid; i.e. not going
  to collide with anything or fall off the edge

== References:
- [Read Developing Games With Ruby]
  (https://leanpub.com/developing-games-with-ruby/read)