Refactor `valid_move?` into Grid class
Refactor the `valid_move?` method implementation out of the Robot class
and into the Grid class so it is the responsibility of world the Robot
inhabits to check if the entity's next move will be valid; i.e. the
entity will not fall off table by not going outside the maximum or
minimum dimensions of said table.

== Notes:
- Add the appropriate tests for the Grid's new `valid_move` method