Add Compass class to represent Vector2D heading
Add a new Compass class to represent (or decorator) a compass heading /
direction from a given Transform position and target. This is in order
to represent the different Compass directions within the application,
whilst fully utilising 2D Vector points to represent virtual space,
rather than baking in the results ahead of time (like in the initial
iteration of the application).

This calculates the heading by first getting the differences between the
target and position x & y coordinates, as `px` & `py` respectively. From
this difference the angle rotation around the origin can be calculated
using the arc tangent `atan2` to get the value of the angle in radians.
Then transform that result into an `octant` to match the list of
specified headings, by dividing the angle within the range of `-PI..PI`,
multiplying, adding and getting the modulus with the length of known
headings `8`, in order to get the index value to determine the current
compass heading.

== Notes:
- Currently the Transform object does not yet existing, however it can
  be simplify described as having a `position` & `target` that it is
  facing, which with that information can determine if it is facing
  North, East, west or South

== References:
- [Stack Overflow - How to snap a directional 2d vector to a Compass N
  NE E SE S SW W NW]
  (http://stackoverflow.com/questions/1437790/how-to-snap-a-directional-2d-vector-to-a-compass-n-ne-e-se-s-sw-w-nw)

- [GameDev - What's the best way of transforming a 2d vector into the
  closest 8 way compass direction]
  (http://gamedev.stackexchange.com/questions/49290/whats-the-best-way-of-transforming-a-2d-vector-into-the-closest-8-way-compass-d)