Update Robot class with Compass for `facing`
Update the Robot class with the new Compass class in order to determine
the current facing direction, as represent on a Compass; i.e. North,
East, South and West. This is in order to later replace the internals of
the Robot class' `@position` and `@facing` instances variables with a
more generic object called Transform, in order to perform transform and
translations of objects without each of them implementing their own
version of Vector2D or coordinates, as well as manipulations of them.

== Notes:
- Add a temporary Transform object to represent the Transform object
  that has yet to be added, by needed by the Compass object to determine
  its current heading

- Tests continue to pass as before