Update Robot with new Transform
Update the Robot class with the new Transform class to represent its
internal `position` and facing direction `target` used for determine its
compass heading. Since the compass heading can now be calculated from
the Transform's internals the `DIRECTIONS` map can be reduce to the
Vector2D that is required in order to move that direction (most of which
are heavily borrowed from the Vector2D class methods).

This _simplifies_ the Robot behaviour through the use of a more generic
system of manipulating its internal `transform` object that can be later
used to compare its position with that of other objects within a similar
system; like boundaries, floor tiles, or table coordinates.

== Notes:
- Update `facing` variable shadowed within the `place` method in order
  to avoid potential conflicts with the class internal `facing` method

- Currently the Transform object returns new transform on `translate`
  and `rotate` methods, this is to prevent side-effects during testing
  and general use, since a copy would need to be made for such a method
  like `next_move` in order to evaluate the next move which moving the
  Robot before it is valid (or once for every `valid_move` call)