Update Transform to mutate state
Update the Transform object to mutate its current position and target
(implied rotation) rather than treating itself like value object, expect
in arithmetic operations; like the `+` operator. This is in order
demonstrate something other than Value Object for the Transform, but
encapsulation of state within an Object and mutating it from the outside
with its provide methods.

Also, in terms of Game / Simulation when there are often several
hunderds (if not thousands) of transforms or similar position
coordinates, replace each one with a new instance of its self would be
quite taxing to perform per frame, so mutating the bare minimum of
variables of an already instantiated object has performance benefits
(although likely to unseen in this small application).

== Notes:
- Remove unnecessary comments to cleanup Transform class

- Update Robot class to reflect changes to its internal Transform, so
  that it does not move when checking to see if the next move is valid,
  otherwise it would move multiple times during a valid move check (once
  for each time it is called) and instead create a copy of the current
  transform to preform checks on

  _Alternatively, the simulation could push back an object on collision;
  e.g. a wall could update the game object's position to the equal
  velocity it exerted on the wall in the opposite direction (something
  along the lines of "equal and opposite reaction"); however would
  require process of input to always be valid and putting the onerous
  the collided with objects (in this case the world) to push back the
  robot onto the table and could be seen as wasted effort_