Add Commands to later replace Robot Commands
Add an implementation of each of the Robot Commands, independent of the
Robot class, to support their reuse with other Entities (that have
references to the "World" and internal position with rotation via a
transform). This is to primarily demonstrate the Command Pattern, even
though the initial version did not require it, but is often expected
within these coding tests.

This will will allow for more multiple commands to created on the Robot
object after it has been initialized, which could lead adding new
commands dynamically or remove them, if need be.

== Notes:
- There are currently comments on the base implementation of the Command
  object to build a `CallAndResponse` type command object; sort of like
  a conditional block that stored and executed later

- Update Transform to allow comparisons between it and other
  transform-like objects

== References:
- [death-of-ifs - YouTube]
  (https://www.youtube.com/watch?v=7K1jB7fTmXQ)

- [Too Many Commands - YouTube]
  (https://www.youtube.com/watch?v=Oh7bim886zk)

- [Bob - exercism.io]
  (http://exercism.io/submissions/51c9ba4c424f5db66300003b)