Refactor Directions into a module for reuse
Refactor the Directions Hash into a module for resue between Commands.
This allows for the Commands to reuse the same source for getting the
Vector2D from a current heading (almost like an Enum in other
languages). Alternatively, this cloud have been included via add the
Module's Constant to the Command's own class or both Movement related
commands could inherit from a common `MovementCommand` that include the
DIRECTIONS constant, however this way helps to ensure there is one
constant source of truth for directions, as well as making it extendable
independent of the classes including it.