Update Robot class to begin to use Command objects
Update the Robot class to begin to use Command objects in place of its
regular methods, by creating a method to process received input from
outside the class and then execute one of its known commands should they
match the input, otherwise preform no action if there are no matches.

Since now a lot of the logic related to the commands is happening within
the Command, some constants and helper methods can be removed from the
Robot class; e.g. `valid_move?`, `next_move`, `facing`, etc; this is
also reflected in the tests for those methods. (This breaks the internal
API, so if being versioned by Semantic Versioning standards would
increase by one major release, but is ignored for the sake of this
coding test).

== Notes:
- At the moment old commands need to be retained, before the Input
  classes can be updated to pass the input as a message to processed
  within the Robot class, however this had been delayed to prevent too
  many changes occurring in one commit, diffusing what is happending and
  making it harder / confusing to understand what changed

- Move any relevant tests from Robot into Command tests

== References:
- [death-of-ifs - YouTube]
  (https://www.youtube.com/watch?v=7K1jB7fTmXQ)

- [Too Many Commands - YouTube]
  (https://www.youtube.com/watch?v=Oh7bim886zk)

- [Bob - exercism.io]
  (http://exercism.io/submissions/51c9ba4c424f5db66300003b)