Add PlayerController class for Input & Robot
Add the Player Controller class to add and in-between class for
processing Input and then handling Entity manipulation; i.e. moving the
Robot. This is in order to add some separation between the Abilities of
a Robot (or other Entities) and the inputs from the outside User (or
Player), since it will then allow for more User control without needing
to go through the Robot every time.

== Notes:
- The criticism of stripping a Robot of its abilities means that a Robot
  will effectively be a dumb data container for Transform manipulation
  :(

- Add appropriate tests for Player Controller (mostly duplicates of
  the Robot tests that will be later stripped)