Add Player Controller to App script
Add the new Player Controller to the application script `app.rb` and
remove the Input class from Robot, passing it to the Controller instead.
This requires a lot of rewiring of internals, since the PlayerController
now has access to both the `scene`/`world`/`grid`/`table` and the robot
meaning it is able to send commands to both of them, based of the
received input this also reduces the role of Input to just parsing
commands, which later could mean the introduction of other Input
sources; e.g. A.I.; that manipulate other entities.

== Notes:
- Update all respective tests to ensure that they still run and remove
  those that are no longer relevant

- This change pushes most of the complexity of handling input to command
  to the controller that could be later adapted to lazy bind commands
  between internal and external names for abilities performed in
  application

- At the moment there is only one entity/actor that has its input being
  handled and is still being used to update the application. This should
  changed so move than one can change in future

- The attribute `scene` was added to the `app.rb` so that it can be
  later referenced for the purposed `draw` command that this is all
  leading up to