Add spaceship operator `<=>` to Vector2D
Add the spaceship operator to Vector2D class in order for it to be
sortable from largest Y value to smallest Y value, and smallest X
value to largest X; similar to a Cartesian plane. This will allow
for it to be more easily print to a console/screen, which works from
top to bottom, the reverse of the expected Y axis assumption.

== Notes:
- Add appropriate tests for comparing Vector2Ds