Add `render!` method to Toy Robot application App
Add a render method to the Toy Robot application `App` class so that it
may draw to the console the current state of the Table. However, it uses
an internal variable called `rendering` to toggle its state to ensure
that it is not unnecessarily rendering the current table every command.

== References:
- [Game Loop · Sequencing Patterns · Game Programming Patterns]
  (http://gameprogrammingpatterns.com/game-loop.html#event-loops)