Add Exit command to Toy Robot application
Add an "Exit" command to the Toy Robot application `App` so that input
text `"EXIT"` can be processed, passed to the appropriate Exit command,
and then in turn quit the application, rather than the Player Controller
returning the last processed command then determining if it should quit.
This is to avoid leaking internal or external Input commands throughout
the application and to have the App run method to modify its state with
those commands rather than doing the work of those commands.

== Notes:
- Add some appropriate tests